### 2024-07-30 #7056 @Kevin-Mattheus-Moerman
- [ ] Can you confirm this project includes automated testing? If not these need to be added and clearly linked to in both the documentation and also the readme.

Note I have just added the `query-scope` label, to trigger a scope review by the editorial board. This is because of the relatively small size of the project. During this scope review we'll determine if the functionality offered is extensive enough and if the project conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The scope review should take about 2 weeks to complete.
