### 2024-08-26 783e10a
Update paper.md
