### 2024-08-27 ff6acf6
Update paper.bib
