+++ Paper/paper.bib
@@ -14,11 +14,13 @@
   month     =  nov,
   year      =  2012,
   language  = "en"
+  doi       = {10.1016/j.mri.2012.05.001.}
 }
 
 @MISC{Creators_The_MONAI_Consortium_undated-or,
   title  = "Project {MONAI}",
   author = "{Creators The MONAI Consortium}"
+  doi    =  "10.48550/arXiv.2211.02701}"
 }
 
 @ARTICLE{Perez-Garcia2021-jf,
@@ -35,6 +37,7 @@
   keywords  = "Data augmentation; Deep learning; Medical image computing;
                Preprocessing",
   language  = "en"
+  doi       = "10.1016/j.cmpb.2021.106236"
 }
 
 @ARTICLE{Anderson2021-fp,
@@ -49,6 +52,7 @@
   month     =  feb,
   year      =  2021,
   language  = "en"
+  doi       = "10.1016/j.prro.2021.02.003"
 }
 
 @ARTICLE{Rufenacht2023-as,
@@ -66,6 +70,7 @@
   keywords  = "Auto-segmentation; DICOM; DICOM RT structure sets; DICOM RTSS
                conversion; Deep learning; Radiotherapy",
   language  = "en"
+  doi       = "10.2139/ssrn.4251277"
 }
 
 @ARTICLE{Whybra2023-en,
@@ -79,6 +84,7 @@
   month    =  sep,
   year     =  2023,
   language = "en"
+  doi      = "10.1038/s41598-023-41475-w"
 }
 
 @ARTICLE{Whybra2023-fi,
@@ -90,22 +96,9 @@
             Bhakti and Bakas, Spyridon and Bettinelli, Andrea and Boellaard,
             Ronald and {Others}",
   year   =  2023
+  doi    = "10.1148/radiol.2020191145"
 }
 
-@ARTICLE{Yousefirizi2023-ax,
-  title     = "Federated testing of {AI} techniques: Towards sharing of
-               implementations, not just code",
-  author    = "Yousefirizi, Fereshteh and Klyuzhin, Ivan and Girum, Kibrom and
-               Uribe, Carlos and Buvat, Irene and Rahmim, Arman",
-  journal   = "J. Nucl. Med.",
-  publisher = "Society of Nuclear Medicine",
-  volume    =  64,
-  number    = "supplement 1",
-  pages     = "P1482--P1482",
-  month     =  jun,
-  year      =  2023,
-  language  = "en"
-}
 
 @ARTICLE{Nioche2018-ct,
   title     = "{LIFEx}: A Freeware for Radiomic Feature Calculation in
@@ -123,5 +116,6 @@
   month     =  aug,
   year      =  2018,
   language  = "en"
+  doi       = "10.1158/0008-5472.can-18-0125" 
 }
 
