### 2024-08-27 cd13f6c
Update paper.md
