### 2024-08-28 #7056 @qurit-frizi
@Kevin-Mattheus-Moerman I addressed the queries about the affiliation, and the DOI of the rereferences.
This project includes automated testing.
