### 2024-08-30 3220ff4
Create CONTRIBUTING.md
