+++ CONTRIBUTING.md
@@ -1,0 +1,19 @@
+# Contributing to [Project Name]
+
+Thank you for considering contributing to our project! Here are some guidelines to help you get started.
+
+## How to Report Issues
+- Before reporting an issue, check if it has already been reported.
+- Include detailed information about the issue, including steps to reproduce, environment details, and screenshots if applicable.
+
+## How to Submit Code
+- Fork the repository and create a new branch for your feature or bug fix.
+- Make sure your code follows the project’s coding style and passes all tests.
+- Submit a pull request with a clear description of your changes and the problem they solve.
+
+## Code of Conduct
+- Please be respectful and considerate in all interactions.
+- Review our [Code of Conduct](link to code of conduct if available) for more details.
+
+## Contact
+- If you have questions or need help, reach out to [contact information].
