+++ README.md
@@ -152,6 +152,10 @@ The add_roi method of our RTStruct class has a multitude of optional parameters 
 #To be added
 nifti to rtstruct
 
+## Contributing
+We welcome contributions to this project! Please review our [CONTRIBUTING guidelines](./CONTRIBUTING.md) for more details.
+
+
 ## How to Cite
 If you are incorporating RT-Utils into your projects, kindly include the following citation:
 
