### 2024-08-30 #7056 @qurit-frizi
@Kevin-Mattheus-Moerman CONTRIBUTING.md  was now added to the readme in the development  branch
