### 2024-09-16 #7056 @Kevin-Mattheus-Moerman
@qurit-frizi the above paper compilation error appears to relate to this .bib file error, and in particular your first reference to 3D slicer. Can you check your .bib file? Perhaps the dot at the end of  this doi should be remove? i.e. `doi       = {10.1016/j.mri.2012.05.001.}` should be `doi       = {10.1016/j.mri.2012.05.001}`?

@openjournals/dev is this indeed the error? :point_up:
