### 2024-09-16 #7056 @Kevin-Mattheus-Moerman
@qurit-frizi I have just removed the `query-scope` label as this submission as passed our initial scope check. Apologies that it took longer than usual. I will now proceed to find a handling editor for this submission.
