### 2024-09-16 #7056 @adamltyson
Hi @qurit-frizi, I'll take over the submission from now on.

I won't start the review until the paper compiles properly. When you get the time, could you take a look at @Kevin-Mattheus-Moerman's comment above and see if this fixes the issue.

In the mean time, we may as well get started finding reviewers. Would you be able to recommend appropriate reviewers for this submission? You can use the [list of potential reviewers](https://reviewers.joss.theoj.org/reviewers) to help. Please provide a list, but don't tag them (with @) so they aren't notified unnecessarily.

Thanks!
