+++ Paper/paper.bib
@@ -14,13 +14,13 @@
   month     =  nov,
   year      =  2012,
   language  = "en"
-  doi       = {10.1016/j.mri.2012.05.001.}
+  doi       = "10.1016/j.mri.2012.05.001"
 }
 
 @MISC{Creators_The_MONAI_Consortium_undated-or,
   title  = "Project {MONAI}",
   author = "{Creators The MONAI Consortium}"
-  doi    =  "10.48550/arXiv.2211.02701}"
+  doi    =  "10.48550/arXiv.2211.02701"
 }
 
 @ARTICLE{Perez-Garcia2021-jf,
