### 2024-09-17 #7056 @adamltyson
Hi @qurit-frizi,

The paper building is still failing. Taking a look at the bib file, some parts of each entry seem to be missing a comma at the end of the line (e.g. `language` in the 3D slicer entry). Could you correct these and we'll try again?
