+++ Paper/paper.bib
@@ -13,13 +13,13 @@
   pages     = "1323--1341",
   month     =  nov,
   year      =  2012,
-  language  = "en"
+  language  = "en",
   doi       = "10.1016/j.mri.2012.05.001"
 }
 
 @MISC{Creators_The_MONAI_Consortium_undated-or,
   title  = "Project {MONAI}",
-  author = "{Creators The MONAI Consortium}"
+  author = "{Creators The MONAI Consortium}",
   doi    =  "10.48550/arXiv.2211.02701"
 }
 
@@ -36,7 +36,7 @@
   year      =  2021,
   keywords  = "Data augmentation; Deep learning; Medical image computing;
                Preprocessing",
-  language  = "en"
+  language  = "en",
   doi       = "10.1016/j.cmpb.2021.106236"
 }
 
@@ -51,7 +51,7 @@
   pages     = "226--229",
   month     =  feb,
   year      =  2021,
-  language  = "en"
+  language  = "en",
   doi       = "10.1016/j.prro.2021.02.003"
 }
 
@@ -69,7 +69,7 @@
   year      =  2023,
   keywords  = "Auto-segmentation; DICOM; DICOM RT structure sets; DICOM RTSS
                conversion; Deep learning; Radiotherapy",
-  language  = "en"
+  language  = "en",
   doi       = "10.2139/ssrn.4251277"
 }
 
@@ -83,7 +83,7 @@
   pages    =  14419,
   month    =  sep,
   year     =  2023,
-  language = "en"
+  language = "en",
   doi      = "10.1038/s41598-023-41475-w"
 }
 
@@ -95,7 +95,7 @@
             Schaer, Roger and Apte, Aditya P and Ayotte, Alexandre and Baheti,
             Bhakti and Bakas, Spyridon and Bettinelli, Andrea and Boellaard,
             Ronald and {Others}",
-  year   =  2023
+  year   =  2023,
   doi    = "10.1148/radiol.2020191145"
 }
 
@@ -115,7 +115,7 @@
   pages     = "4786--4789",
   month     =  aug,
   year      =  2018,
-  language  = "en"
+  language  = "en",
   doi       = "10.1158/0008-5472.can-18-0125" 
 }
 
