### 2024-09-17 #7056 @xuanxu
Merging [this PR](https://github.com/qurit/rt-utils/pull/119) should fix the error
