### 2024-09-19 19f5b16
Update paper.bib
