### 2024-09-23 #7056 @adamltyson
Hi @qurit-frizi,

There still seems to be issues compiling the paper. Could you ensure that it [compiles locally](https://joss.readthedocs.io/en/latest/paper.html#checking-that-your-paper-compiles) then we can try again here?

Thanks!
