+++ Paper/paper.bib
@@ -1,4 +1,5 @@
 @ARTICLE{Fedorov2012-ax,
+
       title     = "{3D} Slicer as an image computing platform for the Quantitative
                    Imaging Network",
       author    = "Fedorov, Andriy and Beichel, Reinhard and Kalpathy-Cramer,
