+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,9 @@
 name: Draft PDF
-on: [push]
+on: 
+  push:
+    paths:
+          - paper/**
+          - .github/workflows/draft-pdf.yml
 
 jobs:
   paper:
@@ -13,7 +17,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: /Paper/paper.md
+          paper-path: paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
