+++ .github/workflows/draft-pdf.yml
@@ -2,8 +2,8 @@ name: Draft PDF
 on: 
   push:
     paths:
-          - Paper/**
-          - .github/workflows/draft-pdf.yml
+      - paper/**
+      - .github/workflows/draft-pdf.yml
 
 jobs:
   paper:
@@ -12,17 +12,18 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
+        
+      - name: List directory contents
+        run: ls -R
+      
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: paper/paper.md  # Update this to match the location of your paper.md file
+      
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
           name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: paper.pdf
+          path: paper/paper.pdf  # Ensure this matches the output path of your generated PDF
