+++ Paper/paper.bib
@@ -118,3 +118,24 @@
       language  = "en",
       doi       = "10.1158/0008-5472.can-18-0125"
     }
+@article{YOUSEFIRIZI2024101745,
+      title = {From code sharing to sharing of implementations: Advancing reproducible AI development for medical imaging through federated testing},
+      journal = {Journal of Medical Imaging and Radiation Sciences},
+      volume = {55},
+      number = {4},
+      pages = {101745},
+      year = {2024},
+      issn = {1939-8654},
+      doi = {https://doi.org/10.1016/j.jmir.2024.101745},
+      url = {https://www.sciencedirect.com/science/article/pii/S1939865424004764},
+      author = {Fereshteh Yousefirizi and Annudesh Liyanage and Ivan S. Klyuzhin and Arman Rahmim},
+      keywords = {Federated testing, Reproducibility, Artificial intelligence, PET segmentation, Preprocessing, Postprocessing, Deployment environment},
+      abstract = {Background
+The reproducibility crisis in AI research remains a significant concern. While code sharing has been acknowledged as a step toward addressing this issue, our focus extends beyond this paradigm. In this work, we explore “federated testing” as an avenue for advancing reproducible AI research and development especially in medical imaging. Unlike federated learning, where a model is developed and refined on data from different centers, federated testing involves models developed by one team being deployed and evaluated by others, addressing reproducibility across various implementations.
+Methods
+Our study follows an exploratory design aimed at systematically evaluating the sources of discrepancies in shared model execution for medical imaging and outputs on the same input data, independent of generalizability analysis. We distributed the same model code to multiple independent centers, monitoring execution in different runtime environments while considering various real-world scenarios for pre- and post-processing steps. We analyzed deployment infrastructure by comparing the impact of different computational resources (GPU vs. CPU) on model performance. To assess federated testing in AI models for medical imaging, we performed a comparative evaluation across different centers, each with distinct pre- and post-processing steps and deployment environments, specifically targeting AI-driven positron emission tomography (PET) imaging segmentation. More specifically, we studied federated testing for an AI-based model for surrogate total metabolic tumor volume (sTMTV) segmentation in PET imaging: the AI algorithm, trained on maximum intensity projection (MIP) data, segments lymphoma regions and estimates sTMTV.
+Results
+Our study reveals that relying solely on open-source code sharing does not guarantee reproducible results due to variations in code execution, runtime environments, and incomplete input specifications. Deploying the segmentation model on local and virtual GPUs compared to using Docker containers showed no effect on reproducibility. However, significant sources of variability were found in data preparation and pre-/post- processing techniques for PET imaging. These findings underscore the limitations of code sharing alone in achieving consistent and accurate results in federated testing.
+Conclusion
+Achieving consistently precise results in federated testing requires more than just sharing models through open-source code. Comprehensive pipeline sharing, including pre- and post-processing steps, is essential. Cloud-based platforms that automate these processes can streamline AI model testing across diverse locations. Standardizing protocols and sharing complete pipelines can significantly enhance the robustness and reproducibility of AI models.}
+}
