+++ Paper/paper.md
@@ -94,7 +94,7 @@ RT-utils spans a diverse range of technical capabilities such as Creating new RT
 For comparing the effects of different RT-Struct conversion methods, we investigated the RT-utils tool, dcmrtstruct2nii (https://github.com/Sikerdebaard/dcmrtstruct2nii) and the built-in tools from LIFEx[@Nioche2018-ct] and 3D Slicer [@Fedorov2012-ax]. We implemented the conversion technique and conducted a comparison of the NIfTI ground truth files. The level of agreement observed between RT-utils and LIFEx surpasses that of other techniques. The mean absolute errors with respect to RT-utils are shown on sagittal and coronal masks. (Figures 1). The visual inspection of an example of converted masks overlaid on PET scans using different techniques are shown in Figures 1. 
 
 
-![The visual inspection of an example of converted masks overlaid on PET scans using different techniques](src/RT.png)
+![The visual inspection of an example of converted masks overlaid on PET scans using different techniques](../src/RT.png)
 
 
 
