### 2024-10-07 #7056 @adamltyson
Thanks @qurit-frizi, the paper looks good now. Would you mind seeing if you can fix the one outstanding BibTeX issue? If you don't know where the problem lies, let me know and I'll see if someone from JOSS can help.
