### 2024-10-09 7a7c9e7
Update paper.bib
