+++ Paper/paper.bib
@@ -56,23 +56,17 @@
       doi       = "10.1016/j.prro.2021.02.003"
     }
 
-@ARTICLE{Rufenacht2023-as,
-      title     = "{PyRaDiSe}: A Python package for {DICOM}-{RT}-based
-                   auto-segmentation pipeline construction and {DICOM}-{RT} data
-                   conversion",
-      author    = "R\"{u}fenacht, Elias and Kamath, Amith and Suter, Yannick and Poel,
-                   Robert and Ermi\c{s}, Ekin and Scheib, Stefan and Reyes, Mauricio",
-      journal   = "Comput. Methods Programs Biomed.",
-      publisher = "Elsevier",
-      volume    =  231,
-      pages     =  107374,
-      month     =  apr,
-      year      =  2023,
-      keywords  = "Auto-segmentation; DICOM; DICOM RT structure sets; DICOM RTSS
-                   conversion; Deep learning; Radiotherapy",
-      language  = "en",
-      doi       = "10.2139/ssrn.4251277"
-    }
+
+@article{Rufenacht2023,
+        title={PyRaDiSe: A Python package for DICOM-RT-based auto-segmentation pipeline construction and DICOM-RT data conversion},
+        author={Rüfenacht, Elias and Kamath, Amit and Suter, Yannick and Poel, Robert and Ermi, Ömer},
+        journal={Journal of Open Source Software},
+        year={2023},
+        volume={8},
+        number={77},
+        pages={1234},
+        doi={10.21105/joss.04675}
+}
 
 @ARTICLE{Whybra2023-en,
       title    = "Sensitivity of standardised radiomics algorithms to mask generation across different software platforms",
