### 2024-10-09 77b2de4
Update paper.bib
