+++ Paper/paper.bib
@@ -81,18 +81,18 @@
       doi      = "10.1038/s41598-023-41475-w"
     }
 
-@ARTICLE{Zwanenburg2020-ibsi,
-      title     = "The Image Biomarker Standardisation Initiative: Standardised Quantitative Radiomics for High-Throughput Image-based Phenotyping",
-      author    = "Zwanenburg, Alex and Valli{\`e}res, Martin and Abdalah, Mohamed A and Aerts, Hugo J W L and Andrearczyk, Vincent and Apte, Aditya and Ashrafinia, Shima and Bakas, Spyridon and Beukinga, Renger J and Boellaard, Ronald and Bogowicz, Margarita and Boldrini, Laura and Buvat, Ir{\`e}ne and Cook, Gary and Davatzikos, Christos and Depeursinge, Adrien and Desseroit, Marie-Claire and Dunet, Vincent and Echegaray, Santiago and El Naqa, Issam and Fave, Xenia and F{\"o}rster, Mathieu and Goh, Vicky and G{\"o}nen, Mehmet and Gollub, Marc and Grossmann, Patrick and Hakimi, Adam A and Hatt, Mathieu and Isensee, Fabian and Lambin, Philippe and Lee, Jeongjin and Leijenaar, Ralph T H and Liptrot, Matthew G and Losnegard, Are and Maier-Hein, Klaus H and Morin, Olivier and Mountz, James M and Nioche, Christophe and Orlhac, Fanny and Peet, Andrew W and Rao, Arvind and Scherer, Daniel and Schleiermacher, Geraldine and Schmidt-Richberg, Alexander and Schoemaker, Matthew and Shafiq-ul Hassan, Muhammad and Siegel, Eliot L and Silva, Mariana and Sousa, Joana and Traverso, Alberto and De Umbrajkar, Mahesh and Valentine, James W and van Griethuysen, Joost J M and Walsh, Sarah and Welk, Alexandra and Wilhelm, Christoph and Wright, Katia L and Yip, Stephen S F and Zeng, Jin and Fennessy, Fiona M and Mak, Raymond H and Pfaehler, Erik and Baessler, Benjamin",
-      journal   = "Radiology",
-      volume    =  295,
-      number    =  2,
-      pages     = "328--338",
-      month     =  may,
-      year      =  2020,
-      language  = "en",
-      doi       = "10.1148/radiol.2020191145"
-    }
+@article{Zwanenburg2020,
+        title     = {The Image Biomarker Standardisation Initiative: Standardised Quantitative Radiomics for High-Throughput Image-based Phenotyping},
[…]
