+++ Paper/paper.bib
@@ -1,127 +1,113 @@
-@ARTICLE{Fedorov2012-ax,
+@article{Fedorov2012,
+  title     = "{3D} Slicer as an image computing platform for the Quantitative Imaging Network",
+  author    = {Fedorov, Andriy and Beichel, Reinhard and Kalpathy-Cramer, Jayashree and Finet, Julien and Fillion-Robin, Jean-Christophe and Pujol, Sonia and Bauer, Christian and Jennings, Dominique and Fennessy, Fiona and Sonka, Milan and Buatti, John and Aylward, Stephen and Miller, James V and Pieper, Steve and Kikinis, Ron},
+  journal   = {Magn. Reson. Imaging},
+  publisher = {Elsevier},
+  volume    = {30},
+  number    = {9},
+  pages     = {1323--1341},
+  month     = {Nov},
+  year      = {2012},
+  language  = {en},
+  doi       = {10.1016/j.mri.2012.05.001}
+}
 
-      title     = "{3D} Slicer as an image computing platform for the Quantitative
-                   Imaging Network",
-      author    = "Fedorov, Andriy and Beichel, Reinhard and Kalpathy-Cramer,
-                   Jayashree and Finet, Julien and Fillion-Robin, Jean-Christophe
-                   and Pujol, Sonia and Bauer, Christian and Jennings, Dominique and
-                   Fennessy, Fiona and Sonka, Milan and Buatti, John and Aylward,
-                   Stephen and Miller, James V and Pieper, Steve and Kikinis, Ron",
-      journal   = "Magn. Reson. Imaging",
-      publisher = "Elsevier",
-      volume    =  30,
-      number    =  9,
-      pages     = "1323--1341",
-      month     =  nov,
-      year      =  2012,
-      language  = "en",
-      doi       = "10.1016/j.mri.2012.05.001"
-    }
+@misc{Creators_The_MONAI_Consortium,
+  title     = {Project {MONAI}},
+  author    = {{Creators The MONAI Consortium}},
+  doi       = {10.48550/arXiv.2211.02701},
+  year      = {2022},
+  url       = {https://arxiv.org/abs/2211.02701}
+}
 
-@MISC{Creators_The_MONAI_Consortium_undated-or,
-      title  = "Project {MONAI}",
-      author = "{Creators The MONAI Consortium}",
-      doi    =  "10.48550/arXiv.2211.02701"
-    }
+@article{Perez-Garcia2021,
+  title     = "{TorchIO}: A Python library for efficient loading, preprocessing, augmentation and patch-based sampling of medical images in deep learning",
+  author    = {P{\'e}rez-Garc{\'i}a, Fernando and Sparks, Rachel and Ourselin, S{\'e}bastien},
+  journal   = {Comput. Methods Programs Biomed.},
+  publisher = {Elsevier},
+  volume    = {208},
+  pages     = {106236},
+  month     = {Sep},
+  year      = {2021},
+  keywords  = {Data augmentation, Deep learning, Medical image computing, Preprocessing},
+  language  = {en},
+  doi       = {10.1016/j.cmpb.2021.106236}
+}
 
-@ARTICLE{Perez-Garcia2021-jf,
-      title     = "{TorchIO}: A Python library for efficient loading, preprocessing,
-                   augmentation and patch-based sampling of medical images in deep
-                   learning",
-      author    = "P\'{e}rez-Garc\'{i}a, Fernando and Sparks, Rachel and Ourselin, S\'{e}bastien",
-      journal   = "Comput. Methods Programs Biomed.",
-      publisher = "Elsevier",
-      volume    =  208,
-      pages     =  106236,
-      month     =  sep,
[…]
