+++ Paper/paper.bib
@@ -1,40 +1,4 @@
-@article{Fedorov2012,
-  title     = "{3D} Slicer as an image computing platform for the Quantitative Imaging Network",
-  author    = {Fedorov, Andriy and Beichel, Reinhard and Kalpathy-Cramer, Jayashree and Finet, Julien and Fillion-Robin, Jean-Christophe and Pujol, Sonia and Bauer, Christian and Jennings, Dominique and Fennessy, Fiona and Sonka, Milan and Buatti, John and Aylward, Stephen and Miller, James V and Pieper, Steve and Kikinis, Ron},
-  journal   = {Magn. Reson. Imaging},
-  publisher = {Elsevier},
-  volume    = {30},
-  number    = {9},
-  pages     = {1323--1341},
-  month     = {Nov},
-  year      = {2012},
-  language  = {en},
-  doi       = {10.1016/j.mri.2012.05.001}
-}
-
-@misc{Creators_The_MONAI_Consortium,
-  title     = {Project {MONAI}},
-  author    = {{Creators The MONAI Consortium}},
-  doi       = {10.48550/arXiv.2211.02701},
-  year      = {2022},
-  url       = {https://arxiv.org/abs/2211.02701}
-}
-
-@article{Perez-Garcia2021,
-  title     = "{TorchIO}: A Python library for efficient loading, preprocessing, augmentation and patch-based sampling of medical images in deep learning",
-  author    = {P{\'e}rez-Garc{\'i}a, Fernando and Sparks, Rachel and Ourselin, S{\'e}bastien},
-  journal   = {Comput. Methods Programs Biomed.},
-  publisher = {Elsevier},
-  volume    = {208},
-  pages     = {106236},
-  month     = {Sep},
-  year      = {2021},
-  keywords  = {Data augmentation, Deep learning, Medical image computing, Preprocessing},
-  language  = {en},
-  doi       = {10.1016/j.cmpb.2021.106236}
-}
-
-@article{Anderson2021,
+@article{Anderson2021-fp,
   title     = "Simple Python Module for Conversions Between {DICOM} Images and Radiation Therapy Structures, Masks, and Prediction Arrays",
   author    = {Anderson, Brian M and Wahid, Kareem A and Brock, Kristy K},
   journal   = {Pract. Radiat. Oncol.},
@@ -48,44 +12,29 @@
   doi       = {10.1016/j.prro.2021.02.003}
 }
 
-@article{Rufenacht2023,
-  title     = {PyRaDiSe: A Python package for DICOM-RT-based auto-segmentation pipeline construction and DICOM-RT data conversion},
-  author    = {R{\"u}fenacht, Elias and Kamath, Amit and Suter, Yannick and Poel, Robert and Ermi, {\"O}mer},
-  journal   = {Journal of Open Source Software},
-  year      = {2023},
-  volume    = {8},
-  number    = {77},
-  pages     = {1234},
-  doi       = {10.21105/joss.04675}
+@misc{Creators_The_MONAI_Consortium_undated-or,
+  title     = {Project {MONAI}},
+  author    = {{Creators The MONAI Consortium}},
+  doi       = {10.48550/arXiv.2211.02701},
+  year      = {2022},
+  url       = {https://arxiv.org/abs/2211.02701}
 }
 
-@article{Whybra2023,
-  title     = "Sensitivity of standardised radiomics algorithms to mask generation across different software platforms",
-  author    = {Whybra, Philip and Spezi, Emiliano},
-  journal   = {Sci. Rep.},
-  volume    = {13},
-  number    = {1},
-  pages     = {14419},
-  month     = {Sep},
-  year      = {2023},
[…]
