### 2024-10-16 #7361 @editorialbot
[REVIEW]: RT-utils: A Minimal Python Library for RT-struct Manipulation
**Submitting author:** @qurit-frizi (<a href="http://orcid.org/0000-0001-5261-6163">Fereshteh Yousefi Rizi</a>)
**Repository:** https://github.com/qurit/rt-utils
**Branch with paper.md** (empty if default branch): development
**Version:** V1.2.0
