### 2024-11-03 #7361 @pchlap
Hi @adamltyson, @qurit-frizi,

Thanks for having me review the JOSS submission for this great library. I have used RT-Utils often in the past and think its a really useful and important tool within Radiation Oncology research.

I do think a bit of work is required to bring this submission in line with JOSS requirements. I have left my comments specific to each JOSS criteria below. In particular I think the JOSS paper needs to be shortened, omitting certain technical details (moving these to the project documentation) and focussing on the JOSS requirements. I recommend referring to [what a JOSS paper should contain](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) and make amendments as necessary. I'd also recommend referring to other JOSS papers to see what they include/exclude (e.g. [PyMedPhys: A community effort to develop an open, Python-based standard library for medical physics applications](https://joss.theoj.org/papers/10.21105/joss.04555)).

I'm here for any questions or clarification needed. I look forward to reviewing the next iteration of this submission!

## General

> **Contribution and authorship:** Has the submitting author ([@qurit-frizi](https://github.com/qurit-frizi)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Could you provide a breakdown of each author's contributions (where it's not already clear from Git commit history)? @qurit-frizi seems to have made changes only to the `README.md` file and I can not determine Armam Rahmim's GitHub user ID to determine their contribution.

## Documentation

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

It would be useful if example data was provided to easily be able to run the examples. This is optional but would be great for someone looking for a quickstart with the tool. Consider hosting some test data on Zenodo which is downloaded as part of the example. Ideally the example would also include an example mask that can then be used in the add_roi function (and not requiring the user to specific this mask to run the example). The example should be runnable from start to finish without requiring the user to make changes to use their data.

> **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

API decription in documentation is incomplete, some details around the `add_roi` function are included, but other functions which are available and details around their inputs and functionality are missing.

>  **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

No community guidelines for contribution to the open-source library are available.

## Paper

>**Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

I feel the summary assumes knowledge of the medical imaging domain, such as what the DICOM standard is and how structures are represented in the RTSTRUCT modality. Providing some additional details for non-specialist audience would be useful.

>**A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

I would like clarification on the statement ...`integrating auto-segmentation solutions using deep learning in clinical environments is rare due to the lack of open-source frameworks that handle DICOM RT-Structure sets effectively.`
This has not necessarily been the primary reason why clinical translation of DL auto-segmentation is challenging. Could you provide a reference? Furthermore, you then list other packages available to achieve this, but do not describe their shortcomings (why not just use them)?

It's also not clear how this tool '`optimizes workflows`'  as is claimed in the statement of need. This needs further clarification what is meant by this.

>**State of the field:** Do the authors describe how this software compares to other commonly-used packages?

As mentioned above, it isn't clear what this library offers in comparison to others available. The real-world example sections also doesn't sufficiently describe the quality of results between this tool and others tested. Some quantative metrics to confirm this would be good which could be presented in a table or plot. The graphic is Figure 1 doesn't support the claims made in the text (for me at least).

> **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is well written overall, however I feel it contains too much specific detail for a JOSS paper. A lot of this detail would be better suited in the tool documentation, where an example could be provided to better understand a specific functionality. I think the content in the **Available Manipulations**, **Handling of DICOM Headers**, **Incorporating an ROI Mask** and **Practical Applications** would be better included in the documentation.

I also do not recommend describing that the tool is hosted on GitHub with the number of stars, and the installation procedure in the JOSS paper. This can easily be found on the GitHub page and documentation.

The use of the term "module" throughout the paper is confusing. Usually this would refer to a specific module within the tool or library. Consider rewording these parts.

Also note the bullet list in the **Available Manipulations** section isn't rendering properly.

> **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

Consider adding a reference I mentioned above (to support the claim that clinical translation of DL models is rare due lack of open-source DICOM RTSTRUCT conversion tools). Also check citation syntax since some references don't seem to render properly in the text.
