+++ Paper/paper.md
@@ -62,40 +62,38 @@ Our module introduces intuitive techniques for efficient data curation of RT-Str
 In the realm of data science, discretized image formats such as NIfTI, NRRD, and MHA are commonly employed, while radiotherapy workflows heavily rely on the DICOM format, specifically the DICOM RT-Struct. Unlike data science architectures like U-Net, which operate on grid-based data, handling the continuously spaced contour points present in RT-Struct poses a unique challenge. To bridge this gap, accurate data conversion between discrete and continuous spaces becomes crucial when working with clinical DICOM RT-Struct data.
 
 ## Technical Overview 
+# RT-utils Overview
 
-The module stands out in its capability to identify discrete Regions of Interest (ROI) and pinpoint their precise spatial locations within intricate structures. Noteworthy is its adept handling of scenarios where multiple ROI names correspond to the same anatomical structure, ensuring a comprehensive and accurate representation of data.
-One of the primary functionalities of the module involves the sophisticated identification of distinct ROI names within RT-Struct files. It also accommodates cases where multiple ROI names map to the same anatomical structure, ensuring precise and comprehensive data organization. The module further facilitates the conversion of DICOM images and RT-Structures into widely adopted formats, specifically NumPy arrays and SimpleITK Images. These standardized formats serve as optimal inputs for various applications, including deep learning models, image analysis pipelines, and radiomic feature calculations. 
+**RT-utils** streamlines RT-Structure file curation by identifying Regions of Interest (ROIs), handling multiple ROI names for the same structure, and converting DICOM RT-Struct and images into formats like NumPy arrays and SimpleITK images. These standardized outputs enable deep learning, image analysis, and radiomics applications. The toolkit also generates DICOM RT-Struct files from predicted arrays, bridging segmentation models and clinical workflows.
 
-Another noteworthy feature is the module ability to simplify the generation of DICOM RT-Struct from predicted NumPy arrays, commonly serving as the output of semantic segmentation deep learning models. This functionality establishes a direct link between predictive modeling and clinical applications, offering a path from automated segmentation to structured RT data. RT-utils employs geometric operations from the Shapely library to convert DICOM ROI Contours into binary masks. This involves creating Shapely polygons from contours and rasterizing these polygons into binary masks. Leveraging geometric principles and libraries, our method accurately represents contours as masks, offering a valuable tool for efficient processing of RT-Struct data in radiotherapy applications.
+The module addresses challenges in converting between discretized image formats (e.g., NIfTI, MHA) and RT-Struct's contour-based format, leveraging geometric operations for accurate binary masks from contour points.
 
-First released in 2020, RT-utils is openly hosted on GitHub (*Starred >180 times as of July 2024), and on PyPI (https://pypi.org/project/rt-utils/) encouraging collaborative development. The demonstrated utility of RT-utils aligns with the evolving landscape of AI-based approaches in the medical domain. In this section, we elaborate on the practical applications of RT-utils. To install RT-utils, simply execute the command pip install RT-utils. 
+Since its 2020 release, RT-utils has grown steadily, with its GitHub repository (*Starred >180 times*) and PyPI presence facilitating collaborative development. Installation is simple: `pip install RT-utils`.
 
-## Available Manipulations
+## Key Features
 
-Once installed, users can import RTStructBuilder to create a new RT-Struct or load an existing one. Upon accomplishing this, users acquire the capability to execute the following operations:
-o	Create a distinctive ROI within the RT-Struct using a binary mask. Optionally, users can specify the color, name, and description of the RT-Struct.
-o	Retrieve a list of names for all ROIs contained within the RT-Struct.
[…]
