### 2024-11-29 3c4773c
Update CONTRIBUTING.md
