+++ CONTRIBUTING.md
@@ -2,9 +2,26 @@
 
 Thank you for considering contributing to our project! Here are some guidelines to help you get started.
 
+**Bugs**
+1. Open an issue. 
+2. Provide as much context as you can about what you're running into.
+3. Provide project and platform versions.
+   
 ## How to Report Issues
 - Before reporting an issue, check if it has already been reported.
 - Include detailed information about the issue, including steps to reproduce, environment details, and screenshots if applicable.
+
+The recommended method of contributing is as follows:
+1. Create an issue on the [issues page](https://github.com/qurit/rt-utils/issues)
+2. Fork the repository to your own account
+3. Fix issue and push changes to your own fork on GitHub
+4. Create a pull request from your fork (whatever branch you worked on) to the development branch in the main repository.
+
+## Sample Data
+The links to the example data are in the:
+rt-utils/tests/mock_data/
+rt-utils/tests/one_slice_data/
+rt-utils/tests/oriented_data/
 
 ## How to Submit Code
 - Fork the repository and create a new branch for your feature or bug fix.
@@ -16,3 +33,6 @@ Thank you for considering contributing to our project! Here are some guidelines 
 
 ## Contact
 - If you have questions or need help, reach out to yousefi.f@gmail.com.
+## License
+
+RT-utils is subject to the `MIT License`, which can be found in the project's root.
