### 2024-11-30 #7361 @qurit-frizi
We have shortened the manuscript based on @pchlap's feedback.

-Here is the detailed breakdown of each author’s contributions:
Conceptualization: Asim Shrestha, Adam Watkins, Carlos Uribe, and Arman Rahmim (@arahmim1) contributed to formulating the idea and designing the solution.
Implementation: Asim Shrestha and Adam Watkins developed the code and implemented the solution.
Repository Maintenance and Manuscript: Fereshteh Yousefirizi and Carlos Uribe have been actively maintaining the repository, addressing comments, and compiling the manuscript.

- The repository now includes:

- Test data and various test codes in [this directory](https://github.com/qurit/rt-utils/tree/development/tests).
- API documentation covering key functionalities such as:
Creating new RT Structs.
Adding contours to existing RT Structs.
Loading RT Struct contours as masks.
Merging two RT Structs.
Configuring additional parameters.
- Additionally, we have updated the contribution guidelines to facilitate better collaboration.
