### 2024-11-30 #7361 @qurit-frizi
@pchlap @adamltyson We revised the summary in response to the reviewer’s comments.
We expanded on the limitations of existing tools to better highlight the need for RT-utils.
We provided a clearer explanation of how RT-utils optimizes workflows by reducing manual effort, improving compatibility, and standardizing data handling.
