### 2024-11-30 #7361 @qurit-frizi
@pchlap @adamltyson Reply to: "The use of the term 'module' throughout the paper is confusing. Usually this would refer to a specific module within the tool or library. Consider rewording these parts."
 We replaced "Module" with "Library" to accurately represent RT-utils as a standalone software package.
