### 2024-11-30 #7361 @qurit-frizi
Reviewer Comment:

@pchlap @adamltyson Reply to: "I also do not recommend describing that the tool is hosted on GitHub with the number of stars, and the installation procedure in the JOSS paper. This can easily be found on the GitHub page and documentation."

You make an excellent point. Including the number of GitHub stars and installation instructions in the paper is unnecessary and can be distracting. We removed References to GitHub Stars: Eliminate mentions of the number of stars from the paper. We omitted the Installation Instructions: Exclude the installation command and direct readers to the GitHub repository and documentation for setup guidance.
