+++ file_conversion_FYR.py
@@ -1,0 +1,251 @@
+import os
+import pydicom
+import numpy as np
+import SimpleITK as sitk
+from rt_utils import RTStructBuilder
+import csv
+import json
+import dateutil
+
+# Define your DICOM root directory here:
+IMAGE_FOLDER_PATH = "///data"
+
+ATTRIBUTE_FILE_NAME = "attributes.csv"
+HEADERS_FILE_NAME = "headers.json"
+SAVE_JSON = False
+
+def winapi_path(dos_path, encoding=None):
+    # Simplified for non-Windows usage:
+    return os.path.abspath(dos_path)
+
+def bqml_to_suv(dcm_file: pydicom.FileDataset) -> float:
+    '''
+    Calculates the SUV conversion factor from Bq/mL to g/mL using DICOM header info.
+    This simplified version returns only the SUV factor.
+    '''
+    nuclide_dose = dcm_file[0x054, 0x0016][0][0x0018, 0x1074].value  # Injected dose (Bq)
+    weight = dcm_file[0x0010, 0x1030].value  # Patient weight (kg)
+    half_life = float(dcm_file[0x054, 0x0016][0][0x0018, 0x1075].value)  # Half life (s)
+
+    series_time = str(dcm_file[0x0008, 0x0031].value)  # Series time (HHMMSS)
+    series_date = str(dcm_file[0x0008, 0x0021].value)  # Series date (YYYYMMDD)
+    series_dt = dateutil.parser.parse(series_date + ' ' + series_time)
+
+    nuclide_time = str(dcm_file[0x054, 0x0016][0][0x0018, 0x1072].value)  # Injection time
+    nuclide_dt = dateutil.parser.parse(series_date + ' ' + nuclide_time)
+
+    delta_time = (series_dt - nuclide_dt).total_seconds()
+    decay_correction = 2 ** (-1 * delta_time/half_life)
+    suv_factor = (weight * 1000) / (decay_correction * nuclide_dose)
+    return suv_factor
+
+def getDicomHeaders(file):
+    dicomHeaders = file.to_json_dict()
+    # remove pixel data from headers
+    dicomHeaders.pop('7FE00010', None)
+    return dicomHeaders
+
+def get_patient_nifti_dir(dicom_dir):
+    # This function finds the patient's directory and creates a NIFTI folder inside it.
+    # Assuming structure: .../data/patientX/DICOM/...
+    # We want: .../data/patientX/NIFTI/
+    patient_dir = dicom_dir
+    # Move up until we exit DICOM directories
+    # Typically, dicom_dir might look like: /.../data/patientX/DICOM/studyY
+    # One dirname: /.../data/patientX/DICOM
+    # Another dirname: /.../data/patientX
+    patient_dir = os.path.dirname(os.path.dirname(dicom_dir))  # This should now point to patientX directory
+    nifti_dir = os.path.join(patient_dir, "NIFTI")
+    if not os.path.exists(nifti_dir):
+        os.makedirs(nifti_dir)
+    return nifti_dir
+
+def dicomToNifti(file, seriesDir):
+    patientID, modality, studyDate = getattr(file, 'PatientID', None), getattr(file, 'Modality', None), getattr(file, 'StudyDate', None)
+    reader = sitk.ImageSeriesReader()
+    seriesNames = reader.GetGDCMSeriesFileNames(seriesDir)
+    reader.SetFileNames(seriesNames)
+    image = reader.Execute()
+
+    # Convert PET to SUV if needed
+    if modality == 'PT':
+        pet = pydicom.dcmread(seriesNames[0])  # read one image
+        suv_factor = bqml_to_suv(pet)
[…]
