### 2024-12-10 #7361 @qurit-frizi
@adamltyson @pchlap the needed conversion from mask to RT-struct has been added to the rt-utils folder in the development branch
