+++ Paper/paper.bib
@@ -70,7 +70,7 @@
   volume    = {8},
   number    = {77},
   pages     = {1234},
-  doi       = {10.21105/joss.04675}
+  doi       = {10.1016/j.cmpb.2023.107374}
 }
 
 @article{Whybra2023-en,
