### 2024-12-10 #7361 @qurit-frizi
> 10.21105/joss.04675

The issue has been resolved
