### 2024-12-13 #7361 @pchlap
> 👋 @pchlap, @suyashkumar, @xtie97 could you check if you can complete the review checklist now that the package has been updated?

Thanks @adamltyson and @qurit-frizi! I am travelling over the weekend, but I will aim to have complete my review by early next week.
