### 2024-12-24 #7361 @qurit-frizi
> @pchlap @adamltyson Reply to: "As mentioned above, it isn't clear what this library offers in comparison to others available. The real-world example sections also doesn't sufficiently describe the quality of results between this tool and others tested. Some quantative metrics to confirm this would be good which could be presented in a table or plot. The graphic is Figure 1 doesn't support the claims made in the text (for me at least)." We described a real world example for the ground truth conversion from RT-struct to masks using quantitative measure, i.e. mean absolute error and the number of pixels in the Table 1 and Figure 6 of following paper: From code sharing to sharing of implementations: Advancing reproducible AI development for medical imaging through federated testing https://www.sciencedirect.com/science/article/pii/S1939865424004764

@pchlap
Table 1:
![image](https://github.com/user-attachments/assets/003a0526-20de-4c2e-8daf-a33ffdf50e96)
Figure 6:
![image](https://github.com/user-attachments/assets/de34b5d1-ec43-4c27-9cb5-4b075b857ff5)
