### 2024-12-24 #7361 @qurit-frizi
@pchlap I may need more clarification on this comment: "While the functionality documentation is looking good, I feel like it would be useful to ensure all methods which are exposed as part of this library have their functionality (and arguments) documented in some way (in addition to the examples). These could even be documented as part of the Python docstrings. This is definitely a nitpick, so isn't necessarily essential from my perspective. Happy to hear what other reviewers think about this too. "Would you please kindly advise?
