### 2024-12-24 3ca768e
Update paper.md
