### 2024-12-24 #7361 @qurit-frizi
@pchlap the issue with the pdf has been resolved.
