### 2025-01-02 #7361 @suyashkumar
Thank you for the great work, and the useful library! And also thanks to @pchlap for giving things a thorough review as well. Putting together some thoughts and questions here, as I go through the checklist:

## Questions

* How can I reproduce Figure 1, and access the data used to create figure 1? (Appears to be required as per the checklist). Are the data used in the mock data in the `tests` directory? Apologies if I’m missing this in the repo.
    * I assume this isn’t original data on human subjects (e.g. it’s data already published elsewhere), but if it is please let us know so we can ensure it complies with JOSS’ policy.
* In the development branch that I’m reviewing, I see that the automated python testing is removed—is there a reason this was removed? In the absence of automated tests, as per the reviewer guidelines we should have detailed manual steps to test in the readme.
* When testing on GitHub code spaces’ VM, it doesn’t come with `libGL.so` shared object, which means I’m unable to test the project based on the installation instructions provided—it fails. Would it make sense to update the readme to say libGL is a required system dependency (and can be installed on linux with `sudo apt install libgl1`)?
    * After installation, I was able to get the tests to pass.

## Suggestions
* Your GitHub actions CI badge on main branch readme appears to be failing (incorrectly), because the latest run of the “Python application” workflow has failed (even though it was not on the main branch). You can update the badge to be branch specific, by constructing a URL like https://github.com/google/cql/actions/workflows/go_build.yaml/badge.svg?branch=main (I don’t think there’s a way to drop in the current readme branch name automatically, [unfortunately](https://github.com/github/markup/issues/913))
    * I also see that the development branch I’m reviewing doesn’t have the Python App testing anymore—is there a reason this was removed?

## Nitpicks + Minor Comments

Below I’ll comment with some Nitpicks, but just some opinionated points I had that might be worth considering to improve the paper. I don’t think these are necessarily blocking, just worth discussion:

* Re: @pchlap’s comment on all public methods having some standard python docstring documentation, I do agree that is reasonable and would be helpful to future users. (I see you have docs on many, but more completeness and adhering to [PEP-257](https://peps.python.org/pep-0008/#documentation-strings) would be great. I think you can stick with one-liners for really obvious use cases as discussed there, but for more complex functions like [add_roi](https://github.com/qurit/rt-utils/blob/544f5b1deb71c80132304b9a5d338f5fdd8dd77e/rt_utils/rtstruct.py#L29) a PEP-257 multiline docstring may make much more sense for users).
* L39/40: “rare due to the lack of open-source frameworks that handle DICOM RT-Structure effectively”: it’s unclear what “effectively” and “necessary functionalities” are later in line 41. It might help future readers to enumerate one or more items specifically that your library adds in contrast to others out there, that you think are “necessary” functionalities. (For example: maybe, ability to convert RT-structs into NumPy arrays and SimpleITK images industry standards with the best ground-truth accuracy, and/or bi-directional conversions between continuous and discrete spaces—if these are lossless bi-directionally, even better).
    * Later in L45-L51 you mention that to address the “gap” of variation in mask-generation methods you provide some additional features. An additional minor nitpick: I’m not sure how many of those listed specifically address the gap of variation in various mask generation algorithms other than providing a [“new standard” library](https://xkcd.com/927/), but I could just be missing it!
    * Edit: Reading your comments above, maybe citing specifically how your package performs well in conversions from the linked paper would make the most sense here as a separate claim, but I’m not sure this addresses the inter-method disagreement directly but it could just be me.
