### 2025-01-06 #7361 @pchlap
> @pchlap I may need more clarification on this comment: "While the functionality documentation is looking good, I feel like it would be useful to ensure all methods which are exposed as part of this library have their functionality (and arguments) documented in some way (in addition to the examples). These could even be documented as part of the Python docstrings. This is definitely a nitpick, so isn't necessarily essential from my perspective. Happy to hear what other reviewers think about this too. "Would you please kindly advise?

@qurit-frizi You can refer to @suyashkumar's comment related to adding docstrings to public methods. I see this is a nitpick for both of us so not strictly required. But at least a more detailed docstring for the `add_roi` function and any other public function which takes several parameters would be nice to have.

Also, thanks for fixing the references. There is just one more I would recommend updating: The reference to the MONAI arxiv paper. You have the author as "The MONAI Consortium Creators". I would change this to the actual authors listed within the DOI: https://arxiv.org/abs/2211.02701
I had the same reference in a previous JOSS paper of mine and here is how I referenced MONAI:
![image](https://github.com/user-attachments/assets/3759a943-979d-47ad-9229-0b9eea08f1b3)
