### 2025-02-01 #7361 @qurit-frizi
**Responses to Questions:**
How can I reproduce Figure 1, and access the data used to create figure 1? (Appears to be required as per the checklist). Are the data used in the mock data in the tests directory? Apologies if I’m missing this in the repo.
I assume this isn’t original data on human subjects (e.g. it’s data already published elsewhere), but if it is please let us know so we can ensure it complies with JOSS’ policy.
_"We have a linked repository in which we provided the detailed examples of the comparison with other conversion techniques, we ''ve now added the link to this repository to the readme under the name of Examples:

You can find the comparison between RT-utils and other conversion techniques in this public repository: https://github.com/qurit/dicom_nifti_conversion_project"__

In the development branch that I’m reviewing, I see that the automated python testing is removed—is there a reason this was removed? In the absence of automated tests, as per the reviewer guidelines we should have detailed manual steps to test in the readme.

_It's modified._
When testing on GitHub code spaces’ VM, it doesn’t come with libGL.so shared object, which means I’m unable to test the project based on the installation instructions provided—it fails. Would it make sense to update the readme to say libGL is a required system dependency (and can be installed on linux with sudo apt install libgl1)?
After installation, I was able to get the tests to pass.

_It's modified._
