### 2025-02-01 #7361 @qurit-frizi
L39/40: “rare due to the lack of open-source frameworks that handle DICOM RT-Structure effectively”: it’s unclear what “effectively” and “necessary functionalities” are later in line 41. It might help future readers to enumerate one or more items specifically that your library adds in contrast to others out there, that you think are “necessary” functionalities. (For example: maybe, ability to convert RT-structs into NumPy arrays and SimpleITK images industry standards with the best ground-truth accuracy, and/or bi-directional conversions between continuous and discrete spaces—if these are lossless bi-directionally, even better).

These lines are modified to make it more clear.
