### 2025-02-01 #7361 @qurit-frizi
Response to @suyashkumar: Later in L45-L51 you mention that to address the “gap” of variation in mask-generation methods you provide some additional features. An additional minor nitpick: I’m not sure how many of those listed specifically address the gap of variation in various mask generation algorithms other than providing a [“new standard” library](https://xkcd.com/927/), but I could just be missing it!

_These lines are modified to make it more clear._
