### 2025-02-01 18fa860
Update rtstruct.py
