+++ .github/workflows/python-app.yml
@@ -28,6 +28,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install -e .  # Ensure local package installation
     - name: Test with pytest
       run: |
         pytest
