### 2025-02-07 e8b8302
Update rtstruct.py
