+++ rt_utils/rtstruct.py
@@ -1,8 +1,6 @@
 from typing import List, Union
-
 import numpy as np
 from pydicom.dataset import FileDataset
-
 from rt_utils.utils import ROIData
 from . import ds_helper, image_helper
 
@@ -17,13 +15,12 @@ class RTStruct:
         self.ds = ds
         self.frame_of_reference_uid = ds.ReferencedFrameOfReferenceSequence[
             -1
-        ].FrameOfReferenceUID  # Use last strucitured set ROI
+        ].FrameOfReferenceUID  # Use last structured set ROI
 
     def set_series_description(self, description: str):
         """
         Set the series description for the RTStruct dataset
         """
-
         self.ds.SeriesDescription = description
 
     def add_roi(
@@ -37,52 +34,49 @@ class RTStruct:
         roi_generation_algorithm: Union[str, int] = 0,
     ):
         """
-        Add a ROI to the rtstruct given a 3D binary mask for the ROI's at each slice
-        Optionally input a color or name for the ROI
-        If use_pin_hole is set to true, will cut a pinhole through ROI's with holes in them so that they are represented with one contour
-        If approximate_contours is set to False, no approximation will be done when generating contour data, leading to much larger amount of contour data
-        This method updates the internal DICOM structure (RTStruct) by adding
-        a new ROIContourSequence, StructureSetROISequence, and
-        RTROIObservationsSequence entry corresponding to the provided mask.
-    
+        Add a Region of Interest (ROI) to the RTStruct given a 3D binary mask for each slice.
+
+        Optionally input a color or name for the ROI. 
+        If `use_pin_hole` is set to True, attempts to handle ROIs with holes by creating a single continuous contour.
+        If `approximate_contours` is set to False, no approximation is done during contour generation, 
+        potentially resulting in a large amount of contour data.
+
+        This method updates the internal DICOM structure (RTStruct) by adding:
+        - ROIContourSequence
+        - StructureSetROISequence
+        - RTROIObservationsSequence
+
         Parameters
         ----------
         mask : np.ndarray
             3D boolean array indicating the ROI. Its shape must match
             the underlying DICOM series in the third dimension.
         color : str or list of int, optional
-            The color representation for the ROI. This can be a string (e.g., "red")
-            or a list of RGB values (e.g., [255, 0, 0]). Defaults to None.
+            Color representation for the ROI (e.g., "red" or [255, 0, 0]). Defaults to None.
         name : str, optional
-            A name/label for this ROI. Defaults to None.
+            Name/label for the ROI. Defaults to None.
         description : str, optional
-            Longer text describing this ROI. Defaults to an empty string.
+            Longer description of the ROI. Defaults to an empty string.
         use_pin_hole : bool, optional
-            If True, attempts to handle ROIs with holes by creating a pinhole
-            (a single continuous contour). Defaults to False.
+            If True, attempts to create a single continuous contour for ROIs with holes. Defaults to False.
         approximate_contours : bool, optional
-            If False, no approximation is done during contour generation,
-            potentially resulting in very large amounts of contour data.
-            Defaults to True.
+            If False, skips approximation during contour generation, leading to larger contour data. Defaults to True.
         roi_generation_algorithm : str or int, optional
-            Identifier for the algorithm used to generate this ROI. This can be
-            a string describing the algorithm or an integer code. Defaults to 0.
-    
+            Identifier for the algorithm used to generate the ROI. Defaults to 0.
+
         Raises
         ------
         ROIException
             - If the mask is not a 3D boolean array.
             - If the mask's shape does not match the loaded DICOM series dimensions.
[…]
