### 2025-02-09 #7361 @pchlap
@adamltyson I've completed my review/checklist and am happy for you to proceed with paper acceptance (pending merge of the `development` branch into `main` as previously mentioned).

@qurit-frizi thank you for addressing all of my previous comments. I think the paper and the repository are looking great. Congratulations on a fantastic piece of work and a valuable contribution to the open-source radiotherapy/medical imaging community!
