### 2025-02-12 #7361 @suyashkumar
Thanks @qurit-frizi for the updates!

- Thanks for including https://github.com/qurit/dicom_nifti_conversion_project for Figure 1. Question: What was the underlying data used to generate this analysis? Is that accessible so I can run a script to reproduce the results? The reviewer checklist says
  > "If the paper contains original data, data are accessible to the reviewers."
  >  "If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item."

  I wasn't able to repro the figure, but if this data/figure was reported elsewhere and isn't considered an original result (is it being in another repo make it considered "reported elsewhere?"), then I think we are all set. Also apologies if I missed something. @adamltyson feel free to weigh in if we can skip this step.

Other than that, everything else seems to be in order and looks good. Thanks!
