+++ README.md
@@ -162,6 +162,7 @@ The add_roi method of our RTStruct class has a multitude of optional parameters 
 
 ## Examples
 You can find the comparison between RT-utils and other conversion techniques in this public repository: https://github.com/qurit/dicom_nifti_conversion_project 
+The real-world example in the JOSS paper under review was generated using HECKTOR Challenge data (patient “CHGJ015”) with the code provided in the (https://github.com/qurit/dicom_nifti_conversion_project ) GitHub repository.
 
 ## Contributing
 We welcome contributions to this project! Please review our [CONTRIBUTING guidelines](./CONTRIBUTING.md) for more details.
