### 2025-02-13 #7361 @qurit-frizi
> Thanks [@qurit-frizi](https://github.com/qurit-frizi) for the updates!
>
> * Thanks for including https://github.com/qurit/dicom_nifti_conversion_project for Figure 1. Question: What was the underlying data used to generate this analysis? Is that accessible so I can run a script to reproduce the results? The reviewer checklist says
>   > "If the paper contains original data, data are accessible to the reviewers."
>   > "If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item."
>
>
>   I wasn't able to repro the figure, but if this data/figure was reported elsewhere and isn't considered an original result (is it being in another repo make it considered "reported elsewhere?"), then I think we are all set. Also apologies if I missed something. [@adamltyson](https://github.com/adamltyson) feel free to weigh in if we can skip this step.
>
> Other than that, everything else seems to be in order and looks good to me! :) Thanks!

Thank you @suyashkumar for following up. Figure 1 was generated from the publicly available HECKTOR dataset, using a specific patient volume and slice selected for illustrative purposes. We have now added the patient ID to the repository README and cited the HECKTOR Challenge dataset in the main text of the paper. The real-world example in the JOSS paper under review was generated using HECKTOR Challenge data (patient “CHGJ015”) with the code provided in [this repository](https://github.com/qurit/dicom_nifti_conversion_project/tree/main).

If the primary goal is to verify code functionality rather than reproduce the exact same visualization, you can run the same steps on any patient volume in the HECKTOR dataset. The pipeline and analysis scripts should work on any of the provided volumes to produce a similar figure (though the slice number will differ). Please let us know if this is helpful. @adamltyson
