### 2025-02-19 #7361 @qurit-frizi
> Hi [@qurit-frizi](https://github.com/qurit-frizi), just for full reproducibility, would it be possible to link from the readme to a standalone script that can reproduce Fig 1 (or a version of it)?

@adamltyson We have updated the real-world example to include a fully reproducible workflow in RTUTILS_read_world_example_for_PET.ipynb, which provides all the necessary code. The images folder can contain either PET or CT DICOM files; in this example, we illustrate how to work with PET data, including converting voxel values to Standardized Uptake Value (SUV) and saving the results as NIfTI files.

We also detail the transformations needed for using RT-Utils, ensuring that the resulting segmentations align correctly with the reference series.  Finally, we have included an example of the usage of dcmrtstruct2nii in the same notebook.  3D Slicer and LIFEx each provide their own built-in tools or plugins for RT-STRUCT → mask conversion. The patient data used for this example cannot be publicly shared. However, one can still use this code, along with the folder structure described in the notebook, on any DICOM PET dataset.
