### 2025-02-19 #7361 @suyashkumar
Thanks @qurit-frizi for the updates, citation, patient ID, and the details around provenance for Figure 1--I think this will be helpful for reproducibility! I don't have access to the HECKTOR data, so can't fully reproduce on my own, but this is probably okay. Thank you for all the work!

I've completed the review checklist and recommend this manuscript for acceptance!
