### 2025-02-19 #7361 @adamltyson
@qurit-frizi the paper looks good, but a few minor comments:
- There's a warning above about a missing DOI. However, I don't think this paper is cited, so you can just remove it from your .bib file
- There's a missing space before the LIFEx citation in the last paragraph
- The paper includes some text about a joint submission to an AAS journal. If this is the case, please add the DOI. I assume it isn't though, so can you just delete these lines (just above the summary section in the .md file)

Once these are updated, could you take a look at the "Additional Author Tasks After Review is Complete" list and let me know when it's all ticked off? Thanks!
