+++ README.md
@@ -13,7 +13,9 @@
 </p>
  
 ---
- 
+<p align="center">
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14902966.svg)](https://doi.org/10.5281/zenodo.14902966)
+
 RT-Utils is motivated to allow physicians and other users to view the results of segmentation performed on a series of DICOM images. RT-Utils allows you to create or load RT Structs, extract 3d masks from RT Struct ROIs, easily add one or more regions of interest, and save the resulting RT Struct in just a few lines!
 You can also use the RT-Utils for merging two existing RT Structs to one file.
 
@@ -174,7 +176,6 @@ We welcome contributions to this project! Please review our [CONTRIBUTING guidel
 
 
 ## How to Cite
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14902966.svg)](https://doi.org/10.5281/zenodo.14902966)
 
 If you are incorporating RT-Utils into your projects, kindly include the following citation:
 
