### 2025-02-21 #7361 @adamltyson
Hi @qurit-frizi, could you make sure that the title and the author list match in both the paper and the zenodo archive? And could you also confirm the version number?
