### 2025-02-21 #7361 @qurit-frizi
> Hi [@qurit-frizi](https://github.com/qurit-frizi), could you make sure that the title and the author list match in both the paper and the zenodo archive? And could you also confirm the version number?
The title and author list match in both the paper and the zenodo archive. The version number is v1.3.0
