### 2025-02-24 #7361 @qurit-frizi
> Hi [@qurit-frizi](https://github.com/qurit-frizi) is `main` the branch with the latest version of the paper? The bot failed to build the PDF as it couldn't find `development`.

@adamltyson  yes we have now merged everything to the main branch as requested, there is no development branch.
